+++ test/unit/test_file_extension.py
@@ -1,6 +1,11 @@
-from gatekeeper import file_extension
+from roverio import file_extension
 
 
 def test_file_extension():
-    result = file_extension.FileExtension.run('gatekeeper', '.py')
-    assert 'gatekeeper/file_extension.py' in result
+    result = file_extension.FileExtension.search_file_extensions('roverio', '.py')
+    assert 'roverio/file_extension.py' in result
+
+
+def test_file_extension_no_results():
+    result = file_extension.FileExtension.search_file_extensions('roverio', '.txt')
+    assert result == []