+++ test/unit/test_scout.py
@@ -1,6 +1,6 @@
-from gatekeeper import scout
+from roverio import scout
 
 
 def test_scout():
-    result = scout.Scout.run('gatekeeper', 'def main()')
-    assert 'File: gatekeeper/scout.py\nSearch: def main():\nLine: 66\n' in result
+    result = scout.Scout.search_for_string('roverio', 'def main()')
+    assert 'File: roverio/scout.py\nSearch: def main():\nLine: 75\n' in result