+++ test/unit/test_secrets.py
@@ -1,11 +1,11 @@
-from gatekeeper import secrets
+from roverio import secrets
 
 
 def test_secrets_no_gitignore():
-    result = secrets.Secrets.run('gatekeeper', 20)
-    assert 'File: gatekeeper/file_extension.py\nSecret: file_extension_files\nLine: 43\n' in result
+    result = secrets.Secrets.search_for_secrets('roverio', 20)
+    assert 'File: roverio/file_extension.py\nSecret: file_extension_files\nLine: 52\n' in result
 
 
 def test_secrets_gitignore():
-    result = secrets.Secrets.run('./', 20)
+    result = secrets.Secrets.search_for_secrets('./', 20)
     assert 'File: ./setup.py\nSecret: long_description_content_type\nLine: 11\n' in result