+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # CHANGELOG
 
+## v2.0.0 (2021-01-30)
+
+* Changed name from `Gatekeeper` to `Rover IO` since most of the tools dealt with file and directory operations instead of protecting code.
+* Added `Sequential Renamer` tool which sequentially and recursively renames files based on the folder they live in
+* Switched from Travis CI to GitHub Actions
+* Moved printing results to a separate function on all tools so we could properly assert outputs in tests while still maintaining console output for the real tool
+
 ## v1.1.0 (2020-09-15)
 
 * Added unit tests and test coverage
 * Scripts now properly return values
 * Updated documentation
 * Added a Makefile
-* Gatekeeper Scout now prints the entire line your search query is found on instead of simply the search query (closes #4)
+* Rover IO Scout now prints the entire line your search query is found on instead of simply the search query (closes #4)
 * Fixed a bug that could not properly open files based on encoding (closes #5)
 * Additional code refactors to place everything into proper classes and functions
 * Automated releasing via Travis